requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
sqlalchemy
pywin32
psutil
pynput
pycaw
Pillow
httpx
aiofiles
requests
comtypes
bcrypt
